+++ internal/handlers/handlers.go
@@ -17,7 +17,6 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/netzen86/collectmetrics/internal/api"
-	"github.com/netzen86/collectmetrics/internal/logger"
 	"github.com/netzen86/collectmetrics/internal/repositories"
 	"github.com/netzen86/collectmetrics/internal/repositories/db"
 	"github.com/netzen86/collectmetrics/internal/repositories/files"
@@ -212,6 +211,7 @@ func JSONUpdateMMHandle(storage repositories.Repo, filename,
 		var err error
 
 		// читаем тело запроса
+		srvlog.Infoln("READING BODY")
 		readedbytes, err := buf.ReadFrom(r.Body)
 		if err != nil {
 			http.Error(w, fmt.Sprintf("%s %v\n", http.StatusText(400), "error body data reading"), 400)
@@ -536,47 +536,6 @@ func PingDB(dbconstring string) http.HandlerFunc {
 		}
 		w.WriteHeader(http.StatusOK)
 	}
-}
-
-// WithLogging функция для включения логирования запросов
-func WithLogging(h http.Handler) http.Handler {
-	logFn := func(w http.ResponseWriter, r *http.Request) {
-		sugar, err := logger.Logger()
-		if err != nil {
-			http.Error(w, fmt.Sprintf("%v %v\n",
-				http.StatusText(http.StatusInternalServerError),
-				http.StatusInternalServerError), 500)
-			return
-		}
-
-		// функция Now() возвращает текущее время
-		start := time.Now()
-
-		// эндпоинт
-		uri := r.RequestURI
-		// метод запроса
-		method := r.Method
-
-		lw, rd := logger.NewLRW(w)
-
-		// точка, где выполняется хендлер pingHandler
-		h.ServeHTTP(lw, r) // обслуживание оригинального запроса
-
-		// Since возвращает разницу во времени между start
-		// и моментом вызова Since. Таким образом можно посчитать
-		// время выполнения запроса.
-		duration := time.Since(start)
-		// fmt.Println(uri, method, duration)
-		// отправляем сведения о запросе в zap
-		sugar.Infoln(
-			"uri", uri,
-			"method", method,
-			"duration", duration,
-			"status", rd.Status,
-			"size", rd.Size,
-		)
-	}
-	return http.HandlerFunc(logFn)
 }
 
 func BadRequest(w http.ResponseWriter, r *http.Request) {
